vector rotation from vector fields defined on tracer cell centers (#153)
* added requirements.txt, info taken from setup.py

added functionality to vector_calc.py to calculate the rotation of a vector
defined on the tracer cell centers. Before this update, the vector rotation
subroutine only worked for vector fields defined on the cell edges.

updated version number to 1.7.1

* added requirements.txt, info taken from setup.py

added functionality to vector_calc.py to calculate the rotation of a vector
defined on the tracer cell centers. Before this update, the vector rotation
subroutine only worked for vector fields defined on the cell edges.

updated version number to 1.7.1

* deleted spuroius jupyter notebook checkpoint file